correcao query

+++ proj_final_aor_backend/src/main/java/aor/paj/proj_final_aor_backend/bean/NotificationBean.java
@@ -326,7 +326,7 @@ public class NotificationBean implements Serializable {
         return true;
     }
 
-    public int getNumberofNotification(long userId) {
+    public long getNumberofNotification(long userId) {
         return notificationDao.numberOfnotificationsByUserID(userId);
     }
